Add large_table demo to the showcase
