+++ examples/showcase/Cargo.toml
@@ -11,3 +11,4 @@ custom_components = { path = "sub/custom_components" }
 dashboard = { path = "sub/dashboard" }
 fragments = { path = "sub/fragments" }
 game_of_life = { path = "sub/game_of_life" }
+large_table = { path = "sub/large_table" }
