+++ examples/showcase/sub/large_table/Cargo.toml
@@ -4,4 +4,4 @@ version = "0.1.0"
 authors = ["qthree <qthree3@gmail.com>"]
 
 [dependencies]
-yew = { path = "../.." }
+yew = { path = "../../../.." }
